requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
streamlit>=1.40.0
apscheduler>=3.10.4
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pandas>=2.2.3
plotly>=5.24.1
markdown>=3.7
pytest>=8.3.3
loguru>=0.7.2
httpx>=0.27.2
pydantic>=2.9.2
streamlit-extras>=0.4.7
